remove old circuit.witness
