+++ src/lib/int.ts
@@ -2,7 +2,7 @@ import { Circuit, Field, Bool } from '../snarky.js';
 import { AnyConstructor, CircuitValue, prop } from './circuit_value.js';
 import { Types } from '../bindings/mina-transaction/types.js';
 import { HashInput } from './hash.js';
-import { inCheckedComputation } from './proof_system.js';
+import { Provable } from './provable.js';
 
 // external API
 export { UInt32, UInt64, Int64, Sign };
@@ -135,7 +135,7 @@ class UInt64 extends CircuitValue {
 
     y_ = y_.seal();
 
-    let q = Circuit.witness(
+    let q = Provable.witness(
       Field,
       () => new Field(x.toBigInt() / y_.toBigInt())
     );
@@ -478,7 +478,7 @@ class UInt32 extends CircuitValue {
 
     y_ = y_.seal();
 
-    let q = Circuit.witness(
+    let q = Provable.witness(
       Field,
       () => new Field(x.toBigInt() / y_.toBigInt())
     );
@@ -875,7 +875,7 @@ class Int64 extends CircuitValue implements BalanceChange {
     // constant case - just return unchecked value
     if (x.isConstant()) return Int64.fromFieldUnchecked(x);
     // variable case - create a new checked witness and prove consistency with original field
-    let xInt = Circuit.witness(Int64, () => Int64.fromFieldUnchecked(x));
+    let xInt = Provable.witness(Int64, () => Int64.fromFieldUnchecked(x));
     xInt.toField().assertEquals(x); // sign(x) * |x| === x
     return xInt;
   }
