+++ src/lib/provable.ts
@@ -5,10 +5,14 @@
  */
 import { Field, Provable as Provable_, Snarky } from '../snarky.js';
 import { FlexibleProvable } from './circuit_value.js';
+import { Context } from './global-context.js';
 import { inCheckedComputation, snarkContext } from './proof_system.js';
 
 // external API
 export { Provable };
+
+// internal API
+export { memoizationContext, memoizeWitness, getBlindingValue };
 
 // TODO move type declaration here
 type Provable<T> = Provable_<T>;
@@ -85,4 +89,44 @@ function clone<T, S extends FlexibleProvable<T>>(type: S, value: T): T {
   let fields = type.toFields(value);
   let aux = type.toAuxiliary(value);
   return (type as Provable<T>).fromFields(fields, aux);
+}
+
+let memoizationContext = Context.create<{
+  memoized: { fields: Field[]; aux: any[] }[];
+  currentIndex: number;
+  blindingValue: Field;
+}>();
+
+/**
+ * Like Circuit.witness, but memoizes the witness during transaction construction
+ * for reuse by the prover. This is needed to witness non-deterministic values.
+ */
+function memoizeWitness<T>(type: FlexibleProvable<T>, compute: () => T) {
+  return Provable.witness<T>(type as Provable<T>, () => {
+    if (!memoizationContext.has()) return compute();
+    let context = memoizationContext.get();
+    let { memoized, currentIndex } = context;
+    let currentValue = memoized[currentIndex];
+    if (currentValue === undefined) {
+      let value = compute();
+      let fields = type.toFields(value).map((x) => x.toConstant());
+      let aux = type.toAuxiliary(value);
+      currentValue = { fields, aux };
+      memoized[currentIndex] = currentValue;
+    }
+    context.currentIndex += 1;
+    return (type as Provable<T>).fromFields(
+      currentValue.fields,
+      currentValue.aux
+    );
+  });
+}
+
+function getBlindingValue() {
+  if (!memoizationContext.has()) return Field.random();
+  let context = memoizationContext.get();
+  if (context.blindingValue === undefined) {
+    context.blindingValue = Field.random();
+  }
+  return context.blindingValue;
 }
